# openapi.yaml
openapi: 3.0.0
info:
  description: "This is the Slurp Server API. You can use it to trigger actions or configure the product."
  version: 1.0.0
  title: Slurp Server
  contact:
    email: xavier.gourmandin@stack-labs.com
  license:
    name: GNU GPL 3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
paths:
...
  /api/{name}/pause:
    post:
      tags:
        - configure
      summary: Pause the given API
      description: Disable slurping of the given API name
      operationId: pauseApi
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The API slurping has been put on hold
        "404":
          description: No API with the given name
  /api/{name}/unpause:
    post:
      tags:
        - configure
      summary: Unpause the given API
      description: Enable slurping of the given API name
      operationId: unpauseApi
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The API slurping has been reactivated
        "404":
          description: No API with the given name
...
servers:
  - url: https://localhost
...